refactor(core): extract page-size helpers out of list framework
Lift the page-size sentinel + helpers (PAGE_SIZE_ALL, PageSizeSentinel,
isPageSizeSentinel, resolvePageSize, getPageSizeLabel, and the default
page-size option arrays) out of the legacy list.component.types into
shared/components/signal-list/page-size.types so they survive deletion
of the list framework. Repoint the live app-paginator + public-api at
the new home; the dead list.component / list.component.types keep using
them via the new import until D deletes them.
