+++ src/frontend/packages/core/src/public-api.ts
@@ -161,6 +161,7 @@ export { ListViewComponent } from './shared/components/list/list-generics/list-v
 export { TableComponent } from './shared/components/list/list-table/table.component';
 export { EnumerateComponent } from './shared/components/enumerate/enumerate.component';
 export * from './shared/components/list/list.component.types';
+export * from './shared/components/signal-list/page-size.types';
 export * from './shared/components/list/data-sources-controllers/list-data-source-types';
 export { RowState, RowsState, getDefaultRowState } from './shared/components/signal-list/row-state.types';
 export * from './shared/components/list/data-sources-controllers/list-data-source';
